feat: add product frontend
category, SPU queries and associations

+++ app/frontend/types/order.go
@@ -4,12 +4,12 @@ type OrderItem struct {
 	ProductName string
 	Picture     string
 	Qty         uint32
-	Cost        float32
+	Cost        float64
 }
 
 type Order struct {
 	OrderId     string
 	CreatedDate string
-	Cost        float32
+	Cost        float64
 	Items       []OrderItem
 }
